+++ src/components/trivia/TriviaCardContainer/TriviaCard/Incomplete/index.tsx
@@ -1,21 +1,22 @@
 import { useRouter } from 'expo-router';
-import Button from '../../../../../../../../components/styled/Button';
+import Button from '../../../../styled/Button';
 import {
   StyledBox,
   StyledColumn,
   StyledRow,
   StyledText,
-} from '../../../../../../../../components/styled/styles';
+} from '../../../../styled/styles';
 import { Dimensions } from 'react-native';
-import { TriviaLoaderIcon } from '../../../../../../../../../assets/icons/TriviaLoaderIcon';
-import { Participant } from '../../../../../../../../components/trivia/LoadingVersus/Participant';
+import { TriviaLoaderIcon } from '../../../../../../assets/icons/TriviaLoaderIcon';
+import { Participant } from '../../../LoadingVersus/Participant';
 import { rgba } from 'polished';
 import { useTheme } from 'styled-components/native';
-import { LerniTriviaIcon } from '../../../../../../../../../assets/icons/LerniTriviaIcon';
-import { TriviaRadialBackground } from '../../../../../../../../../assets/TriviaCardBackground';
+import { LerniTriviaIcon } from '../../../../../../assets/icons/LerniTriviaIcon';
+import { TriviaRadialBackground } from '../../../../../../assets/TriviaCardBackground';
+import ClockIcon from '../../../../../../assets/icons/ClockIcon';
 import { TriviaCardProps } from '../types';
 
-const ChallengedCard = ({
+const IncompleteCard = ({
   programName,
   user,
   opponent,
@@ -56,13 +57,13 @@ const ChallengedCard = ({
         }}
       >
         <StyledText variant="h2" color="white">
-          {'Fuiste retado a un duelo'}
+          {'Trivia incompleta'}
         </StyledText>
         <StyledText variant="body2" color="gray100">
-          {'Responder esta trivia solo te llevar'}
+          {'Termina la partida.'}
         </StyledText>
-        <StyledText variant="body2" color="gray100" css={{ fontFamily: 'Roboto-Bold' }}>
-          {'10 minutos'}
+        <StyledText variant="body2" color="gray100">
+          {'No pierdas la oportunidad de sumar puntos!'}
         </StyledText>
       </StyledColumn>
       <StyledRow css={{ justifyContent: 'center', alignItems: 'flex-end' }}>
@@ -91,4 +92,4 @@ const ChallengedCard = ({
   );
 };
 
-export default ChallengedCard;
+export default IncompleteCard;
