+++ src/components/trivia/TriviaCardContainer/TriviaCard/index.tsx
